requirements.txt
openai>=1.3.0
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
pypdf>=3.17.0
chromadb>=0.4.15
tiktoken>=0.5.1

